+++ src/pt/mult/mult.jl
@@ -14,14 +14,14 @@ The product `ProcessTensor`.
 """
 mult(x::ProcessTensor, y::ProcessTensor, alg::SVDCompression) = mult(x, y, trunc=alg.trunc, verbosity=alg.verbosity)
 """
-	mult(x::ProcessTensor, y::ProcessTensor, alg::DMRGMultAlgorithm)
+	mult(x::ProcessTensor, y::ProcessTensor, alg::DMRGAlgorithm)
 
-Compute the product of two MPO using a DMRG iterative algorithm (e.g., `DMRGMult1`).
+Compute the product of two MPO using a DMRG iterative algorithm (e.g., `DMRG1`).
 
 # Returns
 The product `ProcessTensor`.
 """
-mult(x::ProcessTensor, y::ProcessTensor, alg::DMRGMultAlgorithm) = iterativemult(x, y, alg)
+mult(x::ProcessTensor, y::ProcessTensor, alg::DMRGAlgorithm) = iterativemult(x, y, alg)
 
 
 """
@@ -36,14 +36,14 @@ Equivalent to `mult!(x, y, trunc=alg.trunc, verbosity=alg.verbosity)`.
 """
 mult!(x::ProcessTensor, y::ProcessTensor, alg::SVDCompression) = mult!(x, y, trunc=alg.trunc, verbosity=alg.verbosity)
 """
-	mult!(x::ProcessTensor, y::ProcessTensor, alg::DMRGMultAlgorithm)
+	mult!(x::ProcessTensor, y::ProcessTensor, alg::DMRGAlgorithm)
 
-Compute the product of two MPO in place using a DMRG iterative algorithm (e.g., `DMRGMult1`), storing the result in `x`.
+Compute the product of two MPO in place using a DMRG iterative algorithm (e.g., `DMRG1`), storing the result in `x`.
 
 # Returns
 `x` itself.
 """
-function mult!(x::ProcessTensor, y::ProcessTensor, alg::DMRGMultAlgorithm)
+function mult!(x::ProcessTensor, y::ProcessTensor, alg::DMRGAlgorithm)
 	r = iterativemult(x, y, alg)
 	return copy!(x, r)
 end