+++ test/adt.jl
@@ -46,9 +46,9 @@ end
 	chi = 20
 	trunc = truncdimcutoff(D=chi, ϵ=1.0e-10)
 	alg1 = SVDCompression(trunc)
-	alg2 = DMRGMult1(trunc, initguess=:svd)
-	alg3 = DMRGMult1(trunc, initguess=:rand, maxiter=10)
-	alg4 = DMRGMult1(trunc, initguess=:pre, maxiter=10)
+	alg2 = DMRG1(trunc, initguess=:svd)
+	alg3 = DMRG1(trunc, initguess=:rand, maxiter=10)
+	alg4 = DMRG1(trunc, initguess=:pre, maxiter=10)
 	algs = [alg1, alg2, alg3, alg4]
 	tol = 1.0e-7
 	for T in (Float64, ComplexF64)