+++ test/ptmodels/freebosons.jl
@@ -59,7 +59,7 @@ end
 
 	adag = bosonadagoperator(d=d)
 	hyb = NonDiagonalHyb(adag)
-	algmult = DMRGMult1(trunc)
+	algmult = DMRG1(trunc)
 	algexpan = OverDeterminedProny(n=20, tol=1.0e-8)
 	alg = XTRGIF(k=5, fast=true, algmult=algmult, algexpan=algexpan)
 
@@ -131,7 +131,7 @@ end
 
 	adag = bosonadagoperator(d=d)
 	hyb = NonDiagonalHyb(adag)
-	algmult = DMRGMult1(trunc)
+	algmult = DMRG1(trunc)
 	algexpan = OverDeterminedProny(n=20, tol=1.0e-8)
 	alg = XTRGIF(k=5, fast=true, algmult=algmult, algexpan=algexpan)