+++ test/ptttiif/influenceoperator.jl
@@ -72,7 +72,7 @@ end
 
 		@test distance(mpo1, mps0) / norm(mps0) < dt
 
-		for algmult in (SVDCompression(D=50), DMRGMult1(trunc=truncdimcutoff(D=50,ϵ=1.0e-6)))
+		for algmult in (SVDCompression(D=50), DMRG1(trunc=truncdimcutoff(D=50,ϵ=1.0e-6)))
 			mps1 = differentialinfluencefunctional(lattice, corr, dt, hyb, WII(), algmult, algexpan=algexpan)
 			_n = norm(mps1)
 			mps2 = differentialinfluencefunctional(lattice, corr, dt, hyb, WI(), algmult, algexpan=algexpan)
@@ -227,7 +227,7 @@ end
 		mps0 = mult!(mps0, mps2_mp, trunc=trunc)
 		mps0 = mult!(mps0, mps2_mm, trunc=trunc)
 
-		for algmult in (SVDCompression(D=50), DMRGMult1(trunc=truncdimcutoff(D=50,ϵ=1.0e-6)))
+		for algmult in (SVDCompression(D=50), DMRG1(trunc=truncdimcutoff(D=50,ϵ=1.0e-6)))
 			mps1 = differentialinfluencefunctional(lattice, corr, dt, hyb, WII(), algmult, algexpan=algexpan)
 			_n = norm(mps1)
 			mps2 = differentialinfluencefunctional(lattice, corr, dt, hyb, WI(), algmult, algexpan=algexpan)